Refactored equipping, equip set functionality into another service. Code cleanup
