+++ ActionUI/ActionUI.Plugin/DataModels/Global/CharacterEquipmentSet.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ModifAmorphic.Outward.ActionUI.DataModels.Global
+﻿namespace ModifAmorphic.Outward.ActionUI.DataModels.Global
 {
     public enum EquipmentSetTypes
     {
