+++ ActionUI/ActionUI.Plugin/Startup.cs
@@ -10,7 +10,6 @@ using ModifAmorphic.Outward.Extensions;
 using ModifAmorphic.Outward.GameObjectResources;
 using ModifAmorphic.Outward.Logging;
 using ModifAmorphic.Outward.Modules;
-using ModifAmorphic.Outward.Modules.Crafting;
 using ModifAmorphic.Outward.Unity.ActionMenus;
 using ModifAmorphic.Outward.Unity.ActionUI;
 using System;
@@ -48,8 +47,8 @@ namespace ModifAmorphic.Outward.ActionUI
                                                    services.GetService<LevelCoroutines>(),
                                                    confSettings,
                                                    services.GetService<IModifLogger>))
-                .AddSingleton(new GlobalProfileService(ActionUISettings.CharactersProfilesPath, 
-                                                    ActionUISettings.CharactersProfilesPath, 
+                .AddSingleton(new GlobalProfileService(ActionUISettings.CharactersProfilesPath,
+                                                    ActionUISettings.CharactersProfilesPath,
                                                     services.GetService<IModifLogger>));
